refactor: split contact page into subcomponents
