+++ src/components/ContactForm.jsx
@@ -1,0 +1,117 @@
+import { useState } from 'react';
+import ContactFields from './ContactFields';
+
+export default function ContactForm({ onSuccess }) {
+  const [requestAppointment, setRequestAppointment] = useState(false);
+  const [formData, setFormData] = useState({
+    name: '',
+    email: '',
+    phone: '',
+    documentCategory: '',
+    appointmentType: '',
+    date: '',
+    time: '',
+    message: '',
+  });
+  const [errors, setErrors] = useState({});
+
+  const validate = () => {
+    const newErrors = {};
+    if (!formData.name.trim()) {
+      newErrors.name = 'Full name is required.';
+    }
+    if (!formData.email.trim()) {
+      newErrors.email = 'Email address is required.';
+    } else if (!/\S+@\S+\.\S+/.test(formData.email)) {
+      newErrors.email = 'Enter a valid email address.';
+    }
+    if (!formData.documentCategory) {
+      newErrors.documentCategory = 'Document category is required.';
+    }
+    if (!formData.message.trim()) {
+      newErrors.message = 'Message is required.';
+    }
+    if (requestAppointment) {
+      if (!formData.appointmentType) {
+        newErrors.appointmentType = 'Appointment type is required.';
+      }
+      if (!formData.date) {
+        newErrors.date = 'Preferred date is required.';
+      }
+      if (!formData.time) {
+        newErrors.time = 'Preferred time is required.';
+      }
+    }
+    return newErrors;
+  };
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+    if (errors[name]) {
+      setErrors({ ...errors, [name]: '' });
+    }
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const validation = validate();
+    if (Object.keys(validation).length > 0) {
+      setErrors(validation);
+      return;
+    }
+    setErrors({});
+    onSuccess();
+  };
+
+  return (
+    <>
+      <h1 className="text-4xl font-serif font-semibold tracking-wide heading-gradient text-silver">
+        Contact Us
+      </h1>
+      <form onSubmit={handleSubmit} className="space-y-6" noValidate>
+        <ContactFields
+          formData={formData}
+          errors={errors}
+          onChange={handleChange}
+          requestAppointment={requestAppointment}
+          setRequestAppointment={setRequestAppointment}
+          setErrors={setErrors}
+        />
+        <label className="block" htmlFor="message">
+          <span className="mb-1 block text-platinum">Message</span>
+          <textarea
+            id="message"
+            name="message"
+            required
+            rows="4"
+            value={formData.message}
+            onChange={handleChange}
+            aria-invalid={!!errors.message}
+            className="w-full rounded border border-platinum bg-deepgray px-3 py-2 text-white placeholder-platinum focus:border-silver focus:outline-none"
+          />
+          {errors.message && (
+            <p role="alert" className="mt-1 text-sm text-red-500">
+              {errors.message}
+            </p>
+          )}
+        </label>
+        <button
+          type="submit"
+          className="cta-button hover:border-silver hover:text-silver"
+        >
+          Send Message
+        </button>
+      </form>
+      <p className="text-sm italic text-platinum">
+        Submitting this form does not guarantee an appointment. All requests are
+        subject to confirmation by Keystone Notary Group, LLC.
+      </p>
+      <p className="text-xs text-platinum">
+        Keystone Notary Group, LLC is not a law firm and does not provide legal
+        advice or services. Please consult an attorney for legal questions or
+        document preparation.
+      </p>
+    </>
+  );
+}
