+++ src/components/ThankYou.jsx
@@ -1,0 +1,12 @@
+export default function ThankYou() {
+  return (
+    <>
+      <h1 className="text-4xl font-serif font-semibold tracking-wide heading-gradient text-silver">
+        Thank You
+      </h1>
+      <p className="text-lg text-platinum">
+        Your message has been received. We will be in touch soon.
+      </p>
+    </>
+  );
+}
